requirements.txt
gym
stable-baselines3[extra]
torch
pygame
pillow
requests
aiohttp
numpy
pyboy
aiohttp
python-dotenv